Fixes issue 3279
This was occurring because the user pixel size was larger than the AOI.  In part
this occurred because the user input an AOI that was unprojected but the pixel
size was listed as being in (meters) rather than the projected units of the
AOI polygon.  Updated the UI to indicate it's the units and added a more
helpful error message, AND allowed for floating point values to be passed in
for pixel sizes.


Former-commit-id: e465481b24c33d6a107152fd669cf93e1e059d76

+++ invest_natcap/marine_water_quality/marine_water_quality_biophysical.py
@@ -31,7 +31,7 @@ def execute(args):
             of interest to run the model.  Will define the grid.
         args['land_poly_uri'] (string): OGR polygon DataSource indicating areas
             where land is.
-        args['pixel_size'] (int): float indicating pixel size in meters of
+        args['pixel_size'] (float): float indicating pixel size in meters of
             output grid.
         args['layer_depth'] (float): float indicating the depth of the grid
             cells in meters.
